pendings

+++ src/com/mycompany/myapp/entities/PendingOrders.java
@@ -12,7 +12,7 @@ package com.mycompany.myapp.entities;
 public class PendingOrders {
     private int ID_PendingOrder;
     private User IDUser;
-    private int innonumber;
+    private double innonumber;
     private Oeuvre OeuvreID;
     private int Quantity;
     private String Status;
@@ -20,7 +20,7 @@ public class PendingOrders {
     public PendingOrders() {
     }
 
-    public PendingOrders(int ID_PendingOrder, User IDUser, int innonumber, Oeuvre OeuvreID, int Quantity, String Status) {
+    public PendingOrders(int ID_PendingOrder, User IDUser, double innonumber, Oeuvre OeuvreID, int Quantity, String Status) {
         this.ID_PendingOrder = ID_PendingOrder;
         this.IDUser = IDUser;
         this.innonumber = innonumber;
@@ -45,11 +45,11 @@ public class PendingOrders {
         this.IDUser = IDUser;
     }
 
-    public int getInnonumber() {
+    public double getInnonumber() {
         return innonumber;
     }
 
-    public void setInnonumber(int innonumber) {
+    public void setInnonumber(double innonumber) {
         this.innonumber = innonumber;
     }
